requirements.txt
Django==1.5.4
South==0.8.2
argparse==1.2.1
dj-database-url==0.2.2
django-model-utils==1.5.0
psycopg2==2.5.1
wsgiref==0.1.2
